requirements.txt
arxiv
openai
python-dotenv
tqdm
